P2P: Fix to add addInterfaceToLocalNetwork on DHCP renew
Route for p2p0 interface is added to netd daemon during P2P
connection as part of CMD_POST_DHCP_ACTION. If route is already
added it will throw excpetion. This exception can be ignore as the
route is already added.
During renew case it is failing here as route is added during DHCP
request, hence changing the exception type

Change-Id: I2e644e79e343155640299f4c122e645f88e44312