Wifi: Filter the scan results based on the configured band
Framework needs to get the scan results based on the configured band as
the supplicant / host driver shall maintain the cache for these scan
results. Without this, the UI shall display the scan results not
corresponding to the configured band. This band capability also
needs to be considered while getting the visibility of the network
to trigger auto join. This commit ensures the same.

Change-Id: I6277480eb27e0e115ba4708a454e37d2c9c559a3