WiFi: Check for the status of startMonitoring before enabling p2p interface
Framework shall initialize wlan0 interface followed by p2p0, once the
connection to the supplicant is successful after the completion of wlan0
initialization. This connection to the supplicant is done by invoking
startMonitoring function during both the wlan0 and p2p0 initialization.
If supplicant is terminated before the p2p0 start is triggered (after wlan0
initialization), the control socket's with the supplicant shall be
closed. This shall not trigger the stop of the p2p initialization and any
communication to the supplicant as part of p2p start shall result in to
a fatal exception, for the sockets getting closed. However, the
startMonitoring API as part the p2p0 interface initialization shall know
if the supplicant is terminated/ sockets have got closed. Thus, check if
the status of startMonitoring is success to proceed with further p2p
initializations.

Change-Id: I9da4fefaa0c108b9d634fae2923698c6b76db101